TEIIDDES-1181 ModelExtensionDefinitionBuilder Throws Exception When Deleted Models That Have Saved MEDs. Changed builder to make sure model exists before deleting MED-relating problem markers.
git-svn-id: http://anonsvn.jboss.org/repos/tdesigner/trunk@2126 ceac5f5e-c25f-0410-8b95-ec8c2776e4a6

+++ plugins/org.teiid.designer.extension.ui/src/org/teiid/designer/extension/ui/ModelExtensionDefinitionBuilder.java
@@ -135,8 +135,10 @@ public final class ModelExtensionDefinitionBuilder extends IncrementalProjectBui
                 monitor.subTask(NLS.bind(Messages.modelBuildSubTaskName, modelFile.getName()));
 
                 try {
-                    modelFile.deleteMarkers(ExtensionIds.PROBLEM_MARKER, true, IResource.DEPTH_INFINITE); // clear markers
-                    refreshModelFileMarkers(modelFile); // create MED-related problem markers
+                    if (modelFile.exists()) {
+                        modelFile.deleteMarkers(ExtensionIds.PROBLEM_MARKER, true, IResource.DEPTH_INFINITE); // clear markers
+                        refreshModelFileMarkers(modelFile); // create MED-related problem markers
+                    }
                 } catch (Exception e) {
                     IStatus modelStatus = new Status(IStatus.ERROR, PLUGIN_ID, NLS.bind(Messages.modelFileBuildErrorMsg,
                                                                                         modelFile.getName()), e);
